# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CCBT Dashboard API",
    "description": "FastAPI backend for the CCBT crypto trading bot dashboard (v2).",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/bots/{symbol}": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Bot Detail",
        "description": "Detail for a single bot.",
        "operationId": "bot_detail_api_bots__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trades": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "List Trades",
        "description": "List closed trades, optionally filtered by symbol.",
        "operationId": "list_trades_api_trades_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/equity": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Equity Curve",
        "description": "Equity curve data points.",
        "operationId": "equity_curve_api_equity_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/daily-pnl": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Daily Pnl",
        "description": "Daily PnL aggregated by UTC date.",
        "operationId": "daily_pnl_api_daily_pnl_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyPnlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/logs": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Logs",
        "description": "Recent log lines, optionally filtered.",
        "operationId": "get_logs_api_logs_get",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BotDetailResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "summary": {
            "$ref": "#/components/schemas/BotRow"
          },
          "recent_trades": {
            "items": {
              "$ref": "#/components/schemas/TradeRow"
            },
            "type": "array",
            "title": "Recent Trades"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "summary",
          "recent_trades"
        ],
        "title": "BotDetailResponse",
        "description": "GET /api/bots/{symbol}"
      },
...
      "BotRow": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          },
          "timeframe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeframe"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "position_side": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Side"
          },
          "position_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Size"
          },
          "unrealized_pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unrealized Pnl"
          },
          "total_pnl": {
            "type": "number",
            "title": "Total Pnl"
          },
          "win_rate_pct": {
            "type": "number",
            "title": "Win Rate Pct"
          },
          "profit_factor": {
            "type": "number",
            "title": "Profit Factor"
          },
          "trade_count": {
            "type": "integer",
            "title": "Trade Count"
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "strategy",
          "timeframe",
          "status",
          "position_side",
          "position_size",
          "unrealized_pnl",
          "total_pnl",
          "win_rate_pct",
          "profit_factor",
          "trade_count",
          "last_updated",
          "mode"
        ],
        "title": "BotRow",
        "description": "One row in GET /api/bots."
      },
...
      "DailyPnlPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "pnl": {
            "type": "number",
            "title": "Pnl"
          },
          "trade_count": {
            "type": "integer",
            "title": "Trade Count"
          }
        },
        "type": "object",
        "required": [
          "date",
          "pnl",
          "trade_count"
        ],
        "title": "DailyPnlPoint"
      },
      "DailyPnlResponse": {
        "properties": {
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/DailyPnlPoint"
            },
            "type": "array",
            "title": "Days"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "days"
        ],
        "title": "DailyPnlResponse",
        "description": "GET /api/daily-pnl?symbol="
      },
      "EquityPoint": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "equity": {
            "type": "number",
            "title": "Equity"
          },
          "cumulative_pnl": {
            "type": "number",
            "title": "Cumulative Pnl"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "equity",
          "cumulative_pnl"
        ],
        "title": "EquityPoint"
      },
      "EquityResponse": {
        "properties": {
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/EquityPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "points"
        ],
        "title": "EquityResponse",
        "description": "GET /api/equity?symbol="
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogLine": {
        "properties": {
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "raw": {
            "type": "string",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "level",
          "message",
          "raw"
        ],
        "title": "LogLine"
      },
      "LogsResponse": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LogLine"
            },
            "type": "array",
            "title": "Lines"
          },
          "total_returned": {
            "type": "integer",
            "title": "Total Returned"
          }
        },
        "type": "object",
        "required": [
          "lines",
          "total_returned"
        ],
        "title": "LogsResponse",
        "description": "GET /api/logs?level=&search=&limit="
      },
...
      "TradeListResponse": {
        "properties": {
          "trades": {
            "items": {
              "$ref": "#/components/schemas/TradeRow"
            },
            "type": "array",
            "title": "Trades"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "trades",
          "total"
        ],
        "title": "TradeListResponse",
        "description": "GET /api/trades?symbol=&limit="
      },
      "TradeRow": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "side": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Side"
          },
          "entry_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Price"
          },
          "exit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Price"
          },
          "pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pnl"
          },
          "pnl_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pnl Pct"
          },
          "close_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Reason"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "side",
          "entry_price",
          "exit_price",
          "pnl",
          "pnl_pct",
          "close_reason",
          "timestamp",
          "duration_seconds",
          "strategy"
        ],
        "title": "TradeRow",
        "description": "One closed trade row."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}